% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{methods}
\name{addBatch}
\alias{addBatch}
\alias{addBatch,queue-method}
\title{addBatch}
\usage{
addBatch(object, ...)

\S4method{addBatch}{queue}(object, name = NULL, path = NULL,
  params = NULL, parallelizable = TRUE, waitBeforeNext = TRUE,
  logfile = NULL)
}
\arguments{
\item{object}{An object of class queue.}

\item{...}{Others arguments from specific methods.}

\item{name}{Character Contains alias of the batch. (Default basename(batch@path))}

\item{path}{Character Contains the path to the R batch. (Mandatory)}

\item{params}{Named list Contains the variable to be transfered to batch. (Default NULL)}

\item{parallelizable}{Logical If batch can be launched multiple times at the same moment regardless to groups. (Default TRUE)}

\item{waitBeforeNext}{Logical If queue can launch next batch while this one. (Default TRUE)}

\item{logfile}{Character Contains path to file that contains batch output. (Default queue@logfolder/batch@name.log)}
}
\description{
Add batch to a queue
}
\examples{
\dontrun{
to be seen later
}
}
