% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{methods}
\name{batchFromRank}
\alias{batchFromRank}
\alias{batchFromRank}
\alias{batchFromRank,queue-method}
\title{batchFromRank}
\usage{
batchFromRank(object, Rank)

\S4method{batchFromRank}{queue}(object, Rank)
}
\arguments{
\item{object}{An object of class queue}

\item{Rank}{Character Rank of batch to get its ID}
}
\description{
Get a batch from its Rank

Get a batch from its Rank
}
