% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\docType{class}
\name{classBatch}
\alias{classBatch}
\alias{batch-class}
\title{The batch class}
\description{
This class is used to store follwing slot
}
\section{Slots}{

\describe{
\item{\code{name}}{Character Contains alias of the batch. (Default basename(batch@path))}

\item{\code{desc}}{Character Description. (Optional)}

\item{\code{path}}{Character Contains the path to the R batch. (Mandatory)}

\item{\code{params}}{Named list Contains the variable to be transfered to batch. (Default NULL)}

\item{\code{parallelizable}}{Logical If batch can be launched multiple times at the same moment regardless to groups. (Default TRUE)}

\item{\code{waitBeforeNext}}{Logical If queue can launch next batch while this one. (Default TRUE)}

\item{\code{endIfKO}}{Logical If batch ends KO, forcefully terminate queue. (Default TRUE)}

\item{\code{logfile}}{Character Contains path to file that contains batch output. (Default queue@logfolder/batch@name.log)}

\item{\code{Rank}}{Numeric Contains rank number of batch within the queue. (Not set manually)}
}}

\examples{
showClass("batch")

## Methods available for this class
showMethods(classes="batch")
}
\author{
Quentin Fazilleau
}
