% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{class}
\name{classQueue}
\alias{classQueue}
\alias{queue-class}
\title{The queue class}
\description{
This class is used to store follwing slot
}
\section{Slots}{

\describe{
\item{\code{name}}{Character Contains name of the queue. (Default basename(tempfile(pattern = "file", tmpdir = "")))}

\item{\code{group}}{Character Contains the group name the queue belongs to. (Default NULL)}

\item{\code{folder}}{Character Contains path to a folder that will contains the working directory folder. (Default tempdir())}

\item{\code{batchs}}{List Contains batch S4 objects. (Default NULL)}
}}

\examples{
showClass("queue")

## Methods available for this class
showMethods(classes="queue")
}
\author{
Quentin Fazilleau
}
