% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_func.R
\name{createQueue}
\alias{createQueue}
\title{createQueue}
\usage{
createQueue(name = basename(tempfile(pattern = "queue", tmpdir = "")),
  desc = NULL, group = NULL, owner = Sys.info()["user"], folder = NULL,
  logdir = NULL, clean = TRUE, tmpdir = NULL)
}
\arguments{
\item{name}{Character Alias name of the queue. (Default basename(tempfile(pattern = "file", tmpdir = "")))}

\item{desc}{Character Description}

\item{group}{Character Name of the group the queue belongs to. (Default NULL)}

\item{owner}{Character username of launcher. (Default Sys.info()["user"])}

\item{folder}{Character Path to a folder that will contains the working directory folder. (Default tempdir())}

\item{logdir}{Character Path to the folder that will contains logs file. (By default in folder)}

\item{clean}{Logical Whether or not the working directory folder should be removed. (Default TRUE)}

\item{tmpdir}{Character If use through RSConnect you can redefine a tmpdir not in /tmp/*. (Default NULL)}
}
\description{
Create an empty queue.
}
\examples{
\dontrun{
q <- createQueue()
}
}
