% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_func.R
\name{getHistorizedBatch}
\alias{getHistorizedBatch}
\title{getHistorizedBatch}
\usage{
getHistorizedBatch(reset = FALSE, queueid = NULL, batchid = NULL)
}
\arguments{
\item{reset}{Logical If true then reset table. (Default FALSE)}

\item{queueid}{Numeric vector List of queueid to show. If NULL shows all. (Default NULL)}

\item{batchid}{Numeric vector List of batchid to show. If NULL shows all. (Default NULL)}
}
\description{
Get the waiting queue
}
\examples{
\dontrun{
getHistorizedBatch()
}
}
