% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_func.R
\name{getWaitBatch}
\alias{getWaitBatch}
\title{getWaitBatch}
\usage{
getWaitBatch()
}
\arguments{
\item{reset}{Logical If true then reset table. (Default FALSE)}
}
\description{
Get the waiting queue
}
\examples{
\dontrun{
getWaitBatch()
}
}
