% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_func.R
\name{getWaitQueue}
\alias{getWaitQueue}
\title{getWaitQueue}
\usage{
getWaitQueue(reset = FALSE)
}
\arguments{
\item{reset}{Logical If true then reset table. (Default FALSE)}
}
\description{
Get the waiting queue
}
\examples{
\dontrun{
getWaitQueue()
}
}
