% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{methods}
\name{launch}
\alias{launch}
\alias{launch,queue-method}
\title{launch}
\usage{
launch(object)

\S4method{launch}{queue}(object)
}
\description{
Launch a queue
}
\examples{
\dontrun{
to be seen later
}
}
