% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_func.R
\name{progress}
\alias{progress}
\title{progress}
\usage{
progress(percentage = NULL)
}
\arguments{
\item{percentage}{Numeric Percentage of progress.}
}
\description{
Increase progress of a batch.
}
\examples{
\dontrun{
progress(50)
}
}
