% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_func.R
\name{reset}
\alias{reset}
\title{reset}
\usage{
reset(historized = FALSE)
}
\arguments{
\item{historized}{Logical If historized queue/batchs must be also reset.}
}
\description{
Reset all waiting queue/batchs information.\cr
    This function should be use in case a batch or queue is stuck in the progress batchs.\cr
    That can happen when a process has been killed forcefully.
}
\examples{
\dontrun{
reset()
}
}
