% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{methods}
\name{addBatch}
\alias{addBatch}
\alias{addBatch}
\alias{addBatch,queue-method}
\title{addBatch}
\usage{
addBatch(object, ...)

\S4method{addBatch}{queue}(object, path = NULL, name = NULL,
  params = NULL, parallelizable = TRUE, waitBeforeNext = TRUE,
  logfile = NULL)
}
\arguments{
\item{object}{An object of class queue}

\item{...}{Others arguments from specific methods}

\item{path}{Character Path to the R batch. (Mandatory)}

\item{name}{Character Alias of the batch. (Default basename(batch@path))}

\item{params}{Named list that contains the variable to be transfered to batch. (Default NULL)}

\item{parallelizable}{Logical If batch can be launched multiple times at the same moment regardless to groups. (Default TRUE)}

\item{waitBeforeNext}{Logical If queue can launch next batch while this one. (Default TRUE)}

\item{logfile}{Character Path to file that contains batch output. (Default queue@logfolder/batch@name.log)}
}
\description{
Add batch to a queue
}
\examples{
\dontrun{
q <- createQueue()
q <- addBatch(q, "/path/batch.R")
launch(q)
}
}
\author{
Quentin Fazilleau
}
