% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{class}
\name{classQueue}
\alias{classQueue}
\alias{queue-class}
\title{The queue class}
\description{
This class is used to store follwing slot
}
\section{Slots}{

\describe{
\item{\code{name}}{Character Alias name of the queue. (Default basename(tempfile(pattern = "file", tmpdir = "")))}

\item{\code{desc}}{Character Description. (Optional)}

\item{\code{group}}{Character Name of the group the queue belongs to. (Default NULL)}

\item{\code{owner}}{Character username of launcher. (Default Sys.info()["user"])}

\item{\code{folder}}{Character Path to a folder that will contains the working directory folder. (Default tempdir())}

\item{\code{batchs}}{List Batch S4 objects. (Default NULL)}

\item{\code{logdir}}{Character Path to the folder that will contains logs file. (By default in folder)}

\item{\code{clean}}{Logical Whether or not the working directory folder should be removed. (Default TRUE)}

\item{\code{tmpdir}}{Character If use through RSConnect you can redefine a tmpdir not in /tmp/*. (Default NULL)}
}}

\author{
Quentin Fazilleau
}
