% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_func.R
\name{getHistorizedQueue}
\alias{getHistorizedQueue}
\title{getHistorizedQueue}
\usage{
getHistorizedQueue(queueid = NULL)
}
\arguments{
\item{queueid}{Numeric vector List of queueid to show. If NULL shows all. (Default NULL)}
}
\description{
Get the waiting queue
}
\examples{
\dontrun{
getHistorizedQueue()
}
}
