% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_func.R
\name{launcheR.reset}
\alias{launcheR.reset}
\title{launcheR.reset}
\usage{
launcheR.reset(type = "all", queueid = NULL, prefix = NULL)
}
\arguments{
\item{type}{Character One of c("all", "wait", "historized")}

\item{queueid}{Number vector of queueid(s)}

\item{prefix}{Character prefix of the queuename(s)}
}
\description{
launcheR.reset reset waiting queue/batchs information.\cr
    This function should be use in case a batch or queue is stuck in the progress batchs.\cr
    That can happen when a process has been killed forcefully.
    The function can also be used to remove the historized batchs/queues
}
\examples{
\dontrun{
launcheR.reset()
}
}
